moves isComponentType function to a more appropriate place
